Added subfeature add asset
Subfeature for adding asset to database is added. Several other
configurations also follow, such as addition of @CrossOrigin annotation
for access from Nginx.

+++ model/src/main/java/com/oasis/model/entity/AssetModel.java
@@ -24,7 +24,7 @@ public class AssetModel extends BaseEntity {
     private double price;
 
     @Field(AssetFieldName.ASSET_STOCK)
-    private int stock;
+    private long stock;
 
     @Field(AssetFieldName.ASSET_BRAND)
     private String brand;
